refactor(tags-test): split TagControllerSeamTests.cs for #293 remediation
Tags.Test/TagControllerSeamTests.cs exceeded the 500-line file-size limit
(579 lines), the sole Blocking finding from feature-review
remediation-inputs.2026-07-09T22-52.md. Split into two partial-class files
by logical cohesion: TagControllerSeamTests.cs (dialog-routed methods,
auto-assign flows, property forwarders, shared test-construction helpers,
392 lines) and TagControllerSeamTests.KeyboardNavigation.cs (keyboard and
navigation handlers, 215 lines). Every [TestMethod] and its assertions are
preserved unchanged; no production Tags/*.cs file was modified.

Wired the new file into Tags.Test.csproj (legacy csproj, explicit
<Compile Include>, no glob). Verified via full toolchain re-run: csharpier
clean, analyzer build 0/0, plain Debug build 0/0, all 64 Tags.Test tests
pass, Tags project line coverage unchanged at 92.63% (704/760). Confirmed
via git-stash before/after diff that the solution-wide nullable/
TreatWarningsAsErrors build's 84 pre-existing errors (SVGControl,
UtilitiesSwordfish) are unrelated, out-of-scope vendored debt unaffected
by this change.
